[Common] Add SSSE3 motion compensation routines
The resulting per-block speedup is measured to between ~1.67x and ~2.37x
for horizontal filtering, between ~1.24x and ~1.53x for vertical filtering,
and between ~1.33x and ~2.37x for 2-dimensional filtering on Haswell,
depending on block size, as compared with the existing SSE2 routines.

Outliers are 4-wide vertical and 2-dimensional filtering with a speedups
of ~5.69x and ~6.84x respectively on Haswell (with clang-703.0.31) due to
the C routines used previously.