[Common] Add AVX2 motion compensation routines
The resulting per-block speedup is measured to between ~1.96x and ~3.27x
for horizontal filtering, between ~1.43x and ~2.09x for vertical filtering,
and between ~1.47x and ~3.04x for 2-dimensional filtering on Haswell,
depending on block size, as compared with the existing SSE2 routines.

Outliers are 4-wide vertical and 2-dimensional filtering with a speedups
of ~5.94x and ~8.41x respectively on Haswell (with clang-703.0.31) due to
the C routines used previously.