[Common/x86] Convert McCopyWidthEq8_sse2 back to MMX
Avoid potential performance regressions on CPUs that have lower
throughput for 64-bit loads to xmm registers than to mm registers.

The emms instruction needed with MMX is somewhat costly on recent
Intel cores and causes a slight performance reduction on such cores
as such (hence the motivation for the conversion to SSE).