Simplify getting the iOS SDK path
Since some time ago, the SDK is actually in a directory named
iPhoneOS.sdk, and the one with a version number (like iPhoneOS10.0.sdk)
is only a symbolic link to the other one. Therefore we don't actually
need to query for the current SDK version.

Additionally, xcrun has got another option for getting the full
SDK path, all in one run, instead of having to manually hardcode
the path within the xcode app bundle, and query xcode-select for the
path to the xcode app bundle.