Include -lpthread -lm in the pkg-config file when used as static library
When linking using the C++ frontend, -lm is included automatically,
but it has to be manually specified otherwise.

Likewise the -lpthread flag needs to be added in order to properly
link with the static library; it only worked previously if the user
included that flag already for some other reason.