Add meson build definitions
This is not a complete port, but a good first step towards
using a simpler build system.

In their current state, the build definitions work and
have been tested on Linux and Windows (both against MSVC
and mingw). The build will fail explicitly on all other platforms.

No configuration options are currently exposed.

The test suite has been ported, and the README updated.