Update WelsThreadLib.cpp
cppwinrt headers are included in the windows sdk from 17134 onwards. This gets rid of the C++/CX code in this class and hence can be compiled without the ZW flag, which is very helpful when integrating with other build systems. My personal use case here is building webrtc for uwp.