deblocking: Disambiguate reference pictures
Macroblocks can refer to the same reference picture using
different indices when the same picture appears in multiple
slots in the current reference picture lists. That leads to
calculating the wrong loop filter strength for edge processing,
and small errors in the conformance suite on (for example)
CVWP5_TOSHIBA_E.264 and others.

Modify the test-suite and checksums for affected test files
where this situation occurs.