Adjust the mmi/msa detection mode for mips platform.
Using mips-simd-check.sh to test the current compiler support mmi/msa
or not before make. If supported, enable mmi/msa.
According to the model name in /proc/cpuinfo to test the current
cpu support mmi/msa or not for runtime detection.
Now We can use the following make instructions on mips platform:
1. make (automatic detection mmi/msa)
2. make ENABLE_MMI=No (disable mmi)
3. make ENABLE_MSA=No (disable msa)
4. make ENABLE_MMI=No ENABLE_MSA=No (disable mmi and msa)

Change-Id: Ibd348ebc11912d7fca1b548c76838675d69b7c40
Signed-off-by: gxw <guxiwei-hf@loongson.cn>

Change-Id: I721119c55bdd4d472675e07add17bf8d0ecdbc63