Fix building for arm/aarch64 mingw (with clang)
For aarch64, the assembly can be built as-is with clang, but some
ELF specific directives (that are stripped out by gas-preprocessor
when targeting armasm) need to be ifdeffed out.

For arm, the assembly still needs gas-preprocessor for making it
buildable as thumb, but the same kind of directives must be
ifdeffed out in this case as well.