meson: Respect default_library option
When using library() instead of shared_library() and static_library,
meson will build shared, static, or both depending on the
value of static_library option.

As far as I know extract_all_objects() was uses as workaround for Meson
bugs fixed a while ago when using not installed static libraries.