msvc: Build the C/C++ code with MSVC, not clang-cl, for ARM64
c256ae3b2774030966913e47a9088ec47d053535 switched CCAS to clang-cl.
Previously it was built with gas-preprocessor.pl, which also works,
but is a bit more cumbersome to set up, and clang-cl is distributed
with MSVC these days.

c256ae3b2774030966913e47a9088ec47d053535 also switched CC and CXX
to clang-cl, but that's not strictly needed.

We could of course want to compile all of the code using clang-cl
instead of MSVC's cl.exe itself, but that should probably be a
separate option then, not tied to the architecture.