Save xmm7 in DyadicBilinearQuarterDownsampler_sse
Static analysis of chrome.dll showed that xmm7 was being used but not
preserved in this function. The Windows calling convention requires
that xmm7 be preserved so this change adds the necessary "PUSH_XMM 8"
and POP_XMM directives to fix this.

This fixes issue #3585. This may fix a bug in Chrome but that is
unknown.