Improve checking for GCC >= 8 and Clang for passing -Wno-class-memaccess (#3598)
Don't pass -Wno-class-memaccess to Clang, which doesn't support that
option name. (When inspecting $(CXX) with a pattern, "clang++" also
matches the pattern "%g++".)

Simplify the pattern checking; use $(filter %g++,$(CXX)) instead of
$(patsubst %g++,,$(CXX)) making it less obscure. (This requires
changing the ifeq ($(),) into ifneq.)

Check for %clang++ in addition to %g++.

Also use a similar pattern for checking for clang++ instead of plain
exact matches for "clang++", for passing the option
-Wc++11-compat-reserved-user-defined-literal - this allows applying
the option when CXX contains a path, or if "clang++" is prefixed by
a cross triple.