Use void casts to silence warnings about memcpy to a class
This was previously silenced by passing -Wno-class-memaccess, to
the compiler, if it was supported (for GCC, it was supported since
GCC 8).

Clang supports a similar option, -Wno-nontrivial-memaccess since
Clang 8. It didn't use to warn about these cases, but since Clang 20,
it also warns about this.

Simplify handling the issue by just adding void casts, to avoid
needing to check for whether the options for silencing the warnings
are supported.